+++ lib/array/index.ts
@@ -1,0 +1,27 @@
+'use strict';
+
+import {dedupe}             from './dedupe';
+import {isArray}            from './is';
+import {isNotEmptyArray}    from './isNotEmpty';
+import {join}               from './join';
+import {mapFn}              from './mapFn';
+import {mapKey}             from './mapKey';
+import {mapPrimitive}       from './mapPrimitive';
+import {shuffle}            from './shuffle';
+import {sort}               from './sort';
+
+export {
+    dedupe,
+    isArray,
+    isArray as is,
+    isNotEmptyArray,
+    isNotEmptyArray as isNotEmpty,
+    isNotEmptyArray as isNeArray,
+    isNotEmptyArray as isNe,
+    join,
+    mapFn,
+    mapKey,
+    mapPrimitive,
+    shuffle,
+    sort,
+};